pyproject.toml
[project]
name = "scriba"
version = "0.1.0"
description = "CLI-first document normalization pipeline and evaluation toolkit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cerebras-cloud-sdk>=1.67.0",
    "httpx>=0.27.0",
    "pymupdf4llm>=0.3.4",
    "PyYAML>=6.0.0",
    "tqdm>=4.67.3",
]

[project.scripts]
scriba = "scriba.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scriba"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "markdown>=3.10.2",
    "pillow>=12.1.1",
    "playwright>=1.58.0",
    "pymupdf>=1.27.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=1.3.0",
    "tiktoken>=0.12.0",
]
